+++ CHANGELOG.md
@@ -5,3 +5,13 @@ All notable changes to the "intrexxhelper" extension will be documented in this 
 ## [0.1.0] - 2021-08-05
 
 - Initial release
+
+## [0.1.1] - 2021-08-07
+
+### Changed
+
+- Changed workspace initialization to change language by switchting active text editor
+
+### Fixed
+
+- Fixed $-Symbol in velocity snippets